Remove the cleanup of rules for upgrades as this breaks the kill_block_hack.
Warn loudly on kill_block_hack failure: if it cant remove the rule it means the rule wasnt there in the first place
This is a potential security issue as the user may have had more access than configured. Obviously, this is supposed to
never happen, the check is there just in case.